Added heroButtonLink and refactored
