+++ src/components/TextContainer.tsx
@@ -1,41 +1,17 @@
-import * as React from "react"
-import { useStaticQuery, graphql } from "gatsby"
 import ReactMarkdown from "react-markdown"
 import { FullWidthContainer } from "./Containers"
 import Title from "./Title"
 
-export const TextContainer: React.FC = () => {
-  const data = useStaticQuery(graphql`
-    query {
-      strapiFrontPage {
-        description {
-          data {
-            description
-          }
-        }
-        descriptionTitle
-      }
-    }
-  `)
+interface Props {
+  title: string
+  content: string
+}
 
-  interface StrapiHeroData {
-    description: {
-      data: {
-        description: string
-      }
-    }
-    descriptionTitle: string
-  }
-
-  const { description: descriptionData, descriptionTitle }: StrapiHeroData =
-    data.strapiFrontPage
-
-  const { description } = descriptionData.data
-
+export const TextContainer = ({ title, content }: Props) => {
   return (
     <FullWidthContainer className="relative p-3">
-      <Title>{descriptionTitle}</Title>
-      {description.split("\n\n").map((paragraph: string) => (
+      <Title>{title}</Title>
+      {content.split("\n\n").map((paragraph: string) => (
         <div
           className="font-sans pb-3 text-sm"
           key={paragraph.length + paragraph[0]}
